+++ core/src/main/kotlin/io/violabs/picard/domain/Container.kt
@@ -1,0 +1,13 @@
+package io.violabs.picard.domain
+
+data class Container(
+    val name: String,
+    val image: String,
+    val command: List<String>? = null,
+    val ports: List<Port>? = null
+) {
+
+    data class Port(
+        val containerPort: Int
+    )
+}
