+++ core/src/main/kotlin/io/violabs/picard/domain/Kind.kt
@@ -1,0 +1,13 @@
+package io.violabs.picard.domain
+
+enum class Kind {
+    JOB,
+    KUBELET_CONFIGURATION,
+    POD;
+
+    override fun toString(): String = name
+        .split("_")
+        .joinToString("") {
+            it.lowercase().replaceFirstChar(Char::uppercase)
+        }
+}
