+++ core/src/main/kotlin/io/violabs/picard/domain/Metadata.kt
@@ -1,0 +1,5 @@
+package io.violabs.picard.domain
+
+data class Metadata(
+    val name: String
+)
