+++ core/src/main/kotlin/io/violabs/picard/domain/PodTemplate.kt
@@ -1,0 +1,5 @@
+package io.violabs.picard.domain
+
+data class PodTemplate(
+    val spec: Spec? = null
+)
