+++ core/src/main/kotlin/io/violabs/picard/domain/RestartPolicy.kt
@@ -1,0 +1,9 @@
+package io.violabs.picard.domain
+
+enum class RestartPolicy(val ref: String) {
+    ALWAYS("Always"),
+    ON_FAILURE("OnFailure"),
+    NEVER("Never");
+
+    override fun toString(): String = ref
+}
