+++ core/src/main/kotlin/io/violabs/picard/domain/Spec.kt
@@ -1,0 +1,10 @@
+package io.violabs.picard.domain
+
+/**
+ * @property restartPolicy defaults to Always if not specified when run by kubectl
+ */
+data class Spec(
+    val containers: List<Container>? = null,
+    val restartPolicy: RestartPolicy? = null,
+    val template: PodTemplate? = null
+)
