+++ core/src/main/kotlin/io/violabs/picard/domain/builder/Builder.kt
@@ -1,0 +1,5 @@
+package io.violabs.picard.domain.builder
+
+interface Builder<T> {
+    fun build(): T
+}
