+++ core/src/main/kotlin/io/violabs/picard/domain/builder/ContainersBuilder.kt
@@ -1,0 +1,14 @@
+package io.violabs.picard.domain.builder
+
+import io.violabs.picard.domain.Container
+import io.violabs.picard.scopedBuild
+
+class ContainersBuilder : Builder<List<Container>> {
+    private val containers = mutableListOf<Container>()
+
+    override fun build(): List<Container> = containers
+
+    fun container(scope: ContainerBuilder.() -> Unit) {
+        containers.add(scopedBuild(::ContainerBuilder, scope))
+    }
+}
